Refactor: Remove yarn.lock and update dependencies
This commit removes the outdated `yarn.lock` file and
updates the project dependencies using `pnpm`.  The
`README.md` and `src/commands/verify.ts` files were
updated to reflect changes for JAMHacks 9.  A
`.gitignore` entry was added to ignore the `.idea`
folder.  Finally, the `bot.ts` file was updated to use
`readonly` for the token. (main)